C:\Users\ess5740\Documents\Documents\Personal\Coursera\Practical Machine Learning\Final Project\Coursera Final ML Project ESchnelle Oct_24_2019.Rmd
The goal of this project is to predict the manner in which participants did their exercise.  This report describes our model was built, how cross validation was used, and provides my thoughts on the expected out of sample error, and why I made the choices I did.  After choosing a model, it will be used to predict on the 20 different cases in the Test set.