Add files via upload
R Markdown for project to predict the manner in which participants did their exercise.  This is the "classe" variable in the Training set. Any other variables are predictor factor candidates. This is the R Markdown used to describe the data, build and select the final model.  After choosing a model, this process uses the selected model to predict classe for the 20 different cases in the **Test** set.